+++ src/main/java/si/matjazcerkvenik/alertmonitor/web/uibeans/UiReportBean.java
@@ -22,6 +22,7 @@ import org.primefaces.model.charts.line.LineChartOptions;
 import org.primefaces.model.charts.optionconfig.title.Title;
 import si.matjazcerkvenik.alertmonitor.data.DAO;
 import si.matjazcerkvenik.alertmonitor.model.prometheus.PQueryMessage;
+import si.matjazcerkvenik.alertmonitor.model.prometheus.PQueryResult;
 import si.matjazcerkvenik.alertmonitor.model.prometheus.PrometheusHttpClient;
 import si.matjazcerkvenik.alertmonitor.providers.AbstractDataProvider;
 import si.matjazcerkvenik.alertmonitor.util.AmDateFormat;
@@ -34,6 +35,7 @@ import javax.faces.context.FacesContext;
 import java.io.Serializable;
 import java.text.DecimalFormat;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -49,11 +51,18 @@ public class UiReportBean implements Serializable {
     private AbstractDataProvider adp;
 
     private final String QUERY_PROM_UP_TIME = "round(time()-process_start_time_seconds{job=\"prometheus\"})";
-    private final String QUERY_COUNT_TARGETS_ALL = "count(up)";
-    private final String QUERY_COUNT_TARGETS_DOWN = "count(up==0) + count(probe_success==0)";
-    private final String QUERY_AVERAGE_TARGETS_AVAILABILITY = "(sum(avg_over_time(up[1h])) + sum(avg_over_time(probe_success[1h])) - count(probe_success)) / (count(up) - count(probe_success)) * 100";
-    private final String QUERY_TARGETS_LIVENESS = "(count(up == 1) + count(probe_success == 1) - count(probe_success)) / (count(up) - count(probe_success)) * 100";
-    private final String QUERY_COUNT_JOBS_ALL = "count(count(up) by (job))";
+    private final String QUERY_SCRAPE_INTERVAL_SECONDS = "round((timestamp(up) - timestamp(up offset 1h)) / count_over_time({__name__=~\"up\"}[1h]))";
+    private final String QUERY_TOTAL_TARGETS_COUNT = "count(up)";
+    private final String QUERY_TOTAL_TARGETS_DOWN_COUNT = "count(probe_success==0 OR up==0)";
+    //    private final String QUERY_TOTAL_TARGETS_DOWN_COUNT = "count(up==0) + count(probe_success==0)";
+    private final String QUERY_TOTAL_JOBS_COUNT = "count(count(up) by (job))";
+    private final String QUERY_TARGETS_CURRENT_STATUS = "probe_success OR up";
+    private final String QUERY_TARGETS_DOWNTIME = "(count_over_time(up[7d]) - (sum_over_time(up[7d]))) * __SCRAPE_INTERVAL__";
+    private final String QUERY_TARGETS_AVAILABILITY_PERCENTAGE = "sum_over_time(up[7d]) / count_over_time(up[7d]) * 100";
+    private final String QUERY_TARGETS_MTBF = "sum_over_time(up[30d]) * __SCRAPE_INTERVAL__ / 3600 / resets(up[30d])";
+    private final String QUERY_TARGETS_AVERAGE_AVAILABILITY_PERCENTAGE = "(sum(avg_over_time(up[1h])) + sum(avg_over_time(probe_success[1h])) - count(probe_success)) / (count(up) - count(probe_success)) * 100";
+    private final String QUERY_TARGETS_ALIVENESS_PERCENTAGE = "count(probe_success==1 OR up==1) / count(up) * 100";
+//    private final String QUERY_TARGETS_LIVENESS = "(count(up == 1) + count(probe_success == 1) - count(probe_success)) / (count(up) - count(probe_success)) * 100";
     private final String QUERY_PROMETHEUS_AVERAGE_REQUEST_DURATION_MILLIS_TEMPLATE = "sum(rate(prometheus_http_request_duration_seconds_sum[__INTERVAL__m])) / sum(rate(prometheus_http_request_duration_seconds_count[__INTERVAL__m])) * 1000";
     private final String QUERY_PROMETHEUS_90_PERCENT_REQUEST_DURATION = "histogram_quantile(0.90, sum(rate(prometheus_http_request_duration_seconds_bucket[__INTERVAL__m])) by (le)) * 1000";
     private final String QUERY_PROMETHEUS_AVERAGE_RESPONSE_TIME = "sum(rate(prometheus_http_request_duration_seconds_sum[1m])) / sum(rate(prometheus_http_request_duration_seconds_count[1m])) * 1000";
@@ -64,6 +73,8 @@ public class UiReportBean implements Serializable {
         String id = requestParameterMap.getOrDefault("provider", "null");
         adp = DAO.getInstance().getDataProviderById(id);
         LogFactory.getLogger().info("UiReportBean: init: found provider: " + adp.getProviderConfig().getName() + "@" + adp.getProviderConfig().getUri());
+
+        fillInstanceStatusesTable();
     }
 
     public AbstractDataProvider getAdp() {
@@ -80,15 +91,115 @@ public class UiReportBean implements Serializable {
     }
 
     public String getCountTargetsDown() {
-        PQueryMessage queryMessage = executeQuery(QUERY_COUNT_TARGETS_DOWN);
+        PQueryMessage queryMessage = executeQuery(QUERY_TOTAL_TARGETS_DOWN_COUNT);
         if (queryMessage != null) {
             return queryMessage.getData().getResult().get(0).getValue()[1].toString();
         }
         return "n/a";
     }
 
+    private List<InstanceStatuses> instanceStatusesList;
+
+    public List<InstanceStatuses> getInstanceStatusesTable() {
+        return instanceStatusesList;
+    }
+
+    public void fillInstanceStatusesTable() {
+
+        System.out.println("fillInstanceStatusesTable");
+
+        Map<String, InstanceStatuses> map = new HashMap<>();
+
+        PQueryMessage queryMessage1 = executeQuery(QUERY_TARGETS_CURRENT_STATUS);
+        if (queryMessage1 != null) {
+            for (PQueryResult res : queryMessage1.getData().getResult()) {
+                String instance = res.getMetric().get("instance");
+                String job = res.getMetric().get("job");
+                String value = (String) res.getValue()[1];
+                InstanceStatuses is = map.getOrDefault(instance + job, new InstanceStatuses());
+                is.setInstance(instance);
+                is.setJob(job);
+                is.setStatus(value);
+                map.put(instance + job, is);
+            }
+        }
+
+        PQueryMessage queryMessage2 = executeQuery(QUERY_TARGETS_DOWNTIME.replace("__SCRAPE_INTERVAL__", QUERY_SCRAPE_INTERVAL_SECONDS));
+        if (queryMessage2 != null) {
+            for (PQueryResult res : queryMessage2.getData().getResult()) {
+                String instance = res.getMetric().get("instance");
+                String job = res.getMetric().get("job");
+                Integer value = Integer.parseInt(res.getValue()[1].toString());
+                InstanceStatuses is = map.getOrDefault(instance + job, new InstanceStatuses());
+                is.setInstance(instance);
+                is.setJob(job);
+                is.setDownTime(Formatter.convertToDHMSFormat(value));
+                map.put(instance + job, is);
+            }
+        }
+
+        PQueryMessage queryMessage3 = executeQuery(QUERY_TARGETS_AVAILABILITY_PERCENTAGE);
+        if (queryMessage3 != null) {
+            for (PQueryResult res : queryMessage3.getData().getResult()) {
+                String instance = res.getMetric().get("instance");
+                String job = res.getMetric().get("job");
+                Double value = Double.parseDouble(res.getValue()[1].toString());
+                InstanceStatuses is = map.getOrDefault(instance + job, new InstanceStatuses());
+                is.setInstance(instance);
+                is.setJob(job);
+                is.setAvailability(Formatter.roundDouble(value, 3) + "%");
+                map.put(instance + job, is);
+            }
+        }
+
+        PQueryMessage queryMessage4 = executeQuery(QUERY_TARGETS_MTBF.replace("__SCRAPE_INTERVAL__", QUERY_SCRAPE_INTERVAL_SECONDS));
+        if (queryMessage4 != null) {
+            for (PQueryResult res : queryMessage4.getData().getResult()) {
+                String instance = res.getMetric().get("instance");
+                String job = res.getMetric().get("job");
+                String value = res.getValue()[1].toString();
+                InstanceStatuses is = map.getOrDefault(instance + job, new InstanceStatuses());
+                is.setInstance(instance);
+                is.setJob(job);
+                if (value.equalsIgnoreCase("+Inf")
+                        || value.equalsIgnoreCase("-Inf")
+                        || value.equalsIgnoreCase("NaN")) {
+                    is.setMTBF(value);
+                } else {
+                    Double d = Double.parseDouble(value);
+                    is.setMTBF(Formatter.roundDouble(d, 2) + "h");
+                }
+                map.put(instance + job, is);
+            }
+        }
+
+        PQueryMessage queryMessage5 = executeQuery(QUERY_SCRAPE_INTERVAL_SECONDS);
+        if (queryMessage5 != null) {
+            for (PQueryResult res : queryMessage5.getData().getResult()) {
+                String instance = res.getMetric().get("instance");
+                String job = res.getMetric().get("job");
+                Integer value = Integer.parseInt(res.getValue()[1].toString());
+                InstanceStatuses is = map.getOrDefault(instance + job, new InstanceStatuses());
+                is.setInstance(instance);
+                is.setJob(job);
+                is.setScrapeInterval(value + "s");
+                map.put(instance + job, is);
+            }
+        }
+
+        instanceStatusesList = new ArrayList<>();
+        for (InstanceStatuses i : map.values()) {
+            // ignore instances without status
+            // as some queries return data for last 7 or 30 days, there are instances which don't exist
+            // anymore and therefore have null current status. Typically, containers change their IP addresses
+            // after redeployment, making them completely new instance each time.
+            if (i.getStatus() != null) instanceStatusesList.add(i);
+        }
+
+    }
+
     public String getAverageTargetAvailability() {
-        PQueryMessage queryMessage = executeQuery(QUERY_AVERAGE_TARGETS_AVAILABILITY);
+        PQueryMessage queryMessage = executeQuery(QUERY_TARGETS_AVERAGE_AVAILABILITY_PERCENTAGE);
         if (queryMessage != null) {
             Double d = Double.parseDouble(queryMessage.getData().getResult().get(0).getValue()[1].toString());
             return new DecimalFormat("0.00").format(d);
@@ -97,7 +208,7 @@ public class UiReportBean implements Serializable {
     }
 
     public String getTargetsLiveness() {
-        PQueryMessage queryMessage = executeQuery(QUERY_TARGETS_LIVENESS);
+        PQueryMessage queryMessage = executeQuery(QUERY_TARGETS_ALIVENESS_PERCENTAGE);
         if (queryMessage != null) {
             Double d = Double.parseDouble(queryMessage.getData().getResult().get(0).getValue()[1].toString());
             return new DecimalFormat("0.00").format(d);
@@ -106,7 +217,7 @@ public class UiReportBean implements Serializable {
     }
 
     public String getCountTargetsAll() {
-        PQueryMessage queryMessage = executeQuery(QUERY_COUNT_TARGETS_ALL);
+        PQueryMessage queryMessage = executeQuery(QUERY_TOTAL_TARGETS_COUNT);
         if (queryMessage != null) {
             return queryMessage.getData().getResult().get(0).getValue()[1].toString();
         }
@@ -114,7 +225,7 @@ public class UiReportBean implements Serializable {
     }
 
     public String getCountJobsAll() {
-        PQueryMessage queryMessage = executeQuery(QUERY_COUNT_JOBS_ALL);
+        PQueryMessage queryMessage = executeQuery(QUERY_TOTAL_JOBS_COUNT);
         if (queryMessage != null) {
             return queryMessage.getData().getResult().get(0).getValue()[1].toString();
         }
@@ -181,7 +292,7 @@ public class UiReportBean implements Serializable {
     }
 
     public LineChartModel getAvailabilityLineModel() {
-        PQueryMessage queryMessage = executeQueryRange(QUERY_AVERAGE_TARGETS_AVAILABILITY);
+        PQueryMessage queryMessage = executeQueryRange(QUERY_TARGETS_AVERAGE_AVAILABILITY_PERCENTAGE);
         if (queryMessage != null) {
             availabilityLineModel = createSingleValueLineModel("Availability", availabilityLineModel, queryMessage);
         }
@@ -192,7 +303,7 @@ public class UiReportBean implements Serializable {
     private LineChartModel livenessLineModel;
 
     public LineChartModel getLivenessLineModel() {
-        PQueryMessage queryMessage = executeQueryRange(QUERY_TARGETS_LIVENESS);
+        PQueryMessage queryMessage = executeQueryRange(QUERY_TARGETS_ALIVENESS_PERCENTAGE);
         if (queryMessage != null) {
             livenessLineModel = createSingleValueLineModel("Liveness", livenessLineModel, queryMessage);
         }